change output style

+++ lib/larrow/runner/manager.rb
@@ -16,7 +16,7 @@ module Larrow
 
       def signal_trap
         trap('INT') do
-          RunLogger.color('red').info 'try to release'
+          RunLogger.title 'try to release'
           release
           ::Kernel.exit
         end
@@ -33,27 +33,29 @@ module Larrow
       end
 
       def preload
-        RunLogger.color('yellow').info 'load configuration'
+        RunLogger.title 'load configuration'
         self.vcs.load_configuration
         @state = :preload
       end
 
       def allocate
-        RunLogger.color('yellow').info 'allocating resource'
+        RunLogger.title 'allocate resource'
         begin_at = Time.new
         self.app = Model::App.new vcs
         self.app.assign node: Model::Node.new(*vm.create.first)
         during = sprintf('%.2f', Time.new - begin_at)
-        RunLogger.level(1).color('blue').info "allocated(#{during}s)"
+        RunLogger.level(1).detail "allocated(#{during}s)"
         @state = :allocate
       end
 
       def release
-        RunLogger.color('yellow').info 'releasing resource'
+        RunLogger.title 'release resource'
         begin_at = Time.new
-        app.node.destroy if @state != :release
+        if app.node
+          app.node.destroy if @state != :release
+        end
         during = sprintf('%.2f', Time.new - begin_at)
-        RunLogger.level(1).color('blue').info "released(#{during}s)"
+        RunLogger.level(1).detail "released(#{during}s)"
         @state = :release
       end
     end
